# docs/swagger.yaml
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://github.com/kics223w1/EndpointLab
  description: An alternative to httpbin.org in Gin.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: EndpointLab
  version: "1.0"
paths:
  /anything:
    delete:
      consumes:
      - application/json
      description: Return anything that is passed to the request
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Returns anything that is passed to request
      tags:
      - Anything
    get:
      consumes:
      - application/json
      description: Return anything that is passed to the request
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Returns anything that is passed to request
      tags:
      - Anything
    head:
      consumes:
      - application/json
      description: Return anything that is passed to the request
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Returns anything that is passed to request
      tags:
      - Anything
    options:
      consumes:
      - application/json
      description: Return anything that is passed to the request
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Returns anything that is passed to request
      tags:
      - Anything
    patch:
      consumes:
      - application/json
      description: Return anything that is passed to the request
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Returns anything that is passed to request
      tags:
      - Anything
    post:
      consumes:
      - application/json
      description: Return anything that is passed to the request
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Returns anything that is passed to request
      tags:
      - Anything
    put:
      consumes:
      - application/json
      description: Return anything that is passed to the request
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Returns anything that is passed to request
      tags:
      - Anything
  /cache:
    get:
      parameters:
      - description: Optional header to check if the resource has been modified since
          the specified date
        in: header
        name: If-Modified-Since
        type: string
      - description: Optional header to check if the resource matches the given ETag
        in: header
        name: If-None-Match
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Cached response
        "304":
          description: Modified
      summary: Returns a 304 if an If-Modified-Since header or If-None-Match is present.
        Returns the same as a GET otherwise.
      tags:
      - Response Inspection
  /cache/{value}:
    get:
      parameters:
      - description: Number of seconds for the Cache-Control max-age directive
        in: path
        name: value
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Cache control set
      summary: Sets a Cache-Control header for n seconds.
      tags:
      - Response Inspection
  /cookies:
    get:
      consumes:
      - application/json
      description: Returns the cookies sent by the client
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Get cookies
      tags:
      - Cookies
  /cookies/delete:
    get:
      consumes:
      - application/json
      parameters:
      - description: Freeform cookie value
        in: query
        name: freeform
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Redirects to cookie list
          schema:
            type: string
      summary: Deletes cookie(s) as provided by the query string and redirects to
        cookie list.
      tags:
      - Cookies
  /cookies/set:
    get:
      consumes:
      - application/json
      parameters:
      - description: Freeform cookie value
        in: query
        name: freeform
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Redirects to cookie list
          schema:
            type: string
      summary: Sets cookie(s) as provided by the query string and redirects to cookie
        list.
      tags:
      - Cookies
  /cookies/set/{name}/{value}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Cookie name
        in: path
        name: name
        required: true
        type: string
      - description: Cookie value
        in: path
        name: value
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Redirects to cookie list
          schema:
            type: string
      summary: Set a cookie with specified name and value
      tags:
      - Cookies
  /delete:
    delete:
      consumes:
      - application/json
      description: Returns the DELETE parameters of the request
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: The request's DELETE parameters.
      tags:
      - HTTP Methods
  /etag/{etag}:
    get:
      parameters:
      - description: Optional header to check if the resource does not match the given
          ETag
        in: header
        name: If-None-Match
        type: string
      - description: Optional header to check if the resource matches the given ETag
        in: header
        name: If-Match
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Normal response
        "412":
          description: match
      summary: Assumes the resource has the given etag and responds to If-None-Match
        and If-Match headers appropriately.
      tags:
      - Response Inspection
  /get:
    get:
      consumes:
      - application/json
      description: Returns the query parameters of the request
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: The request's query parameters.
      tags:
      - HTTP Methods
  /headers:
    get:
      consumes:
      - application/json
      description: Returns all headers of the request
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Returns the request headers.
      tags:
      - Request Inspection
  /image:
    get:
      description: Returns a simple image of the type suggested by the Accept header
      produces:
      - image/webp
      - ' image/png'
      - ' image/jpeg'
      - ' image/svg+xml'
      responses:
        "200":
          description: OK
          schema:
            type: file
        "406":
          description: Not Acceptable
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Returns an image based on the Accept header
      tags:
      - Images
  /image/jpeg:
    get:
      description: Returns a simple JPEG image
      produces:
      - image/jpeg
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: Returns a JPEG image
      tags:
      - Images
  /image/png:
    get:
      description: Returns a simple PNG image
      produces:
      - image/png
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: Returns a PNG image
      tags:
      - Images
  /image/svg:
    get:
      description: Returns a simple SVG image
      produces:
      - image/svg+xml
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: Returns an SVG image
      tags:
      - Images
  /image/webp:
    get:
      description: Returns a simple WEBP image
      produces:
      - image/webp
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: Returns a WEBP image
      tags:
      - Images
  /ip:
    get:
      consumes:
      - application/json
      description: Returns the IP address of the client making the request
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Returns the client's IP address.
      tags:
      - Request Inspection
  /patch:
    patch:
      consumes:
      - application/json
      description: Returns the PATCH parameters of the request
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: The request's PATCH parameters.
      tags:
      - HTTP Methods
  /post:
    post:
      consumes:
      - application/json
      description: Returns the POST parameters of the request
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: The request's POST parameters.
      tags:
      - HTTP Methods
  /put:
    put:
      consumes:
      - application/json
      description: Returns the PUT parameters of the request
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: The request's PUT parameters.
      tags:
      - HTTP Methods
  /response-headers:
    get:
      consumes:
      - application/json
      description: Returns all response headers including freeform values.
      parameters:
      - description: Freeform query parameter
        in: query
        name: freeform
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Response headers
      summary: Returns response headers.
      tags:
      - Response Inspection
    post:
      consumes:
      - application/json
      description: Returns all response headers including freeform values.
      parameters:
      - description: Freeform query parameter
        in: query
        name: freeform
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Response headers
      summary: Returns response headers.
      tags:
      - Response Inspection
  /status/{code}:
    delete:
      description: Returns a status code based on the path
      parameters:
      - description: HTTP Status Code
        in: path
        name: code
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Return status code or random status code if more than one are given
      tags:
      - Status
    get:
      description: Returns a status code based on the path
      parameters:
      - description: HTTP Status Code
        in: path
        name: code
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Return status code or random status code if more than one are given
      tags:
      - Status
    head:
      description: Returns a status code based on the path
      parameters:
      - description: HTTP Status Code
        in: path
        name: code
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Return status code or random status code if more than one are given
      tags:
      - Status
    options:
      description: Returns a status code based on the path
      parameters:
      - description: HTTP Status Code
        in: path
        name: code
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Return status code or random status code if more than one are given
      tags:
      - Status
    patch:
      description: Returns a status code based on the path
      parameters:
      - description: HTTP Status Code
        in: path
        name: code
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Return status code or random status code if more than one are given
      tags:
      - Status
    post:
      description: Returns a status code based on the path
      parameters:
      - description: HTTP Status Code
        in: path
        name: code
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Return status code or random status code if more than one are given
      tags:
      - Status
    put:
      description: Returns a status code based on the path
      parameters:
      - description: HTTP Status Code
        in: path
        name: code
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Return status code or random status code if more than one are given
      tags:
      - Status
  /user-agent:
    get:
      consumes:
      - application/json
      description: Returns the User-Agent string of the request
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Returns the User-Agent string.
      tags:
      - Request Inspection
swagger: "2.0"